requirements.txt
-e git+https://github.com/mitsuhiko/flask.git@afd3c4532b8625729bed9ed37a3eddd0b7b3b5a9#egg=Flask-dev
Flask-Assets==0.10
Flask-PyMongo==0.3.0
Flask-Redis==0.0.5
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.3
pymongo==2.7.2
pytest==2.5.2
redis==2.10.1
